Fix timestamp tx synchronous issue in use of FeeProvider
Glueby::Internal::Wallet#broadcast(tx) changes the txid of tx in arguments while using FeeProvider. If it store a txid before broadcasting, the txid would be changed.
Here the method can receive a code block which is called before actual broadcasting and caller can execute some codes with real tx that is going to be broadcasted.